# docs/services/gateway/api/public.openapi.yaml
openapi: 3.0.3
info:
  title: Power Knowledge Platform Gateway API
  version: 0.1.0
  license:
    name: Unspecified
    url: https://github.com/Sakayori-Iroha-168/Software_Teamwork
  description: |
    Public caller-facing API contract for the gateway service.
    All stable public and service-to-service HTTP APIs must use RESTful resource-oriented paths.
    This revision treats gateway health checks, user/session,
    knowledge-owned knowledge base, document upload, processing, content, chunk, and retrieval routes,
    document-owned report generation and report settings routes, qa-owned intelligent question answering
    routes, admin-facing runtime model/parser configuration routes, admin user
    management, personal profile, required password-change routes, and gateway-owned
    app version freshness routes as active contracts. Both admin-overview and
    admin-metrics are active gateway contracts for the administration dashboard.
...
servers:
- url: /
  description: Gateway root
security:
- bearerAuth: []
...
paths:
...
  /api/v1/users:
    post:
      tags:
      - auth
      summary: Create user
      operationId: createUser
      security: []
      x-owner-service: auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User and session created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          $ref: '#/components/responses/Error'
  /api/v1/sessions:
    post:
      tags:
      - auth
      summary: Create session
      operationId: createSession
      security: []
      x-owner-service: auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '200':
          description: Session created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /api/v1/admin/users:
    get:
      tags:
      - admin-users
      summary: List manageable users
      operationId: listAdminUsers
      x-owner-service: auth
      x-gateway-role: Public admin entrypoint, admin authorization, response envelope, and auth context forwarding. Auth owns list scoping, filters, roles, and account state.
      description: |
        Lists users the caller is allowed to manage. Administrators see
        standard users only. Super administrators see standard and admin users.
        A bare system:admin permission without the admin or super_admin role
        does not grant user-management authority. Super administrator accounts are
        excluded from this management list. Filtering and pagination are
        performed by Gateway/Auth, not by frontend-only filtering.
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      - name: username
        in: query
        required: false
        description: Case-insensitive username substring filter.
        schema:
          type: string
          minLength: 1
          maxLength: 128
      - name: role
        in: query
        required: false
        description: Managed role filter. Super administrator rows are never returned.
        schema:
          type: string
          enum:
          - standard
          - admin
      - name: status
        in: query
        required: false
        description: Account status filter.
        schema:
          $ref: '#/components/schemas/UserStatus'
      responses:
        '200':
          description: Manageable users.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserListResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
    post:
      tags:
      - admin-users
      summary: Create managed user
      operationId: createAdminUser
      x-owner-service: auth
      x-gateway-role: Public admin entrypoint, admin authorization, response envelope, and auth context forwarding. Auth owns user creation, temporary password hashing, role assignment, and must-change-password state.
      description: |
        Creates a user managed by administrators. This is distinct from public
        self-service registration at POST /api/v1/users: no session is returned
        to the administrator, and the created user must change the temporary
        password on first login.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAdminUserRequest'
      responses:
        '201':
          description: Managed user created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Opaque bearer access token issued by auth and cached by gateway through a token hash; callers must not treat it as a JWT.
  parameters:
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimited:
      description: |
        Gateway or owner-service rate limit. Retry-After may be present when a
        downstream owner service provides a known wait time; gateway-local
        process saturation normally omits it.
      headers:
        Retry-After:
          description: Seconds to wait before retrying when the wait time is known.
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      required:
      - code
      - message
      - requestId
      properties:
        code:
          type: string
          enum:
          - validation_error
          - unauthorized
          - forbidden
          - not_found
          - conflict
          - rate_limited
          - dependency_error
          - internal_error
          - not_implemented
          - unsupported_intent
          - unsupported_mode
        message:
          type: string
        requestId:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
    CreateUserRequest:
      type: object
      required:
      - username
      - password
      description: |
        Public self-service registration request. This route creates a standard
        user, returns a session, and does not set mustChangePassword. It does
        not collect displayName, email, or phone; those are edited later through
        the current-user profile resource.
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 128
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 1024
          writeOnly: true
    CreateSessionRequest:
      type: object
      required:
      - username
      - password
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    UserSummary:
      type: object
      required:
      - id
      - username
      - roles
      - permissions
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: string
          description: Optional display name. May be empty; clients may fall back to username.
        email:
          type: string
          format: email
          nullable: true
          description: Optional non-unique profile email. Not a login identifier.
        phone:
          type: string
          nullable: true
          description: Optional non-unique profile phone. Not a login identifier.
        status:
          $ref: '#/components/schemas/UserStatus'
        mustChangePassword:
          type: boolean
          description: True when the authenticated user must complete the required password-change flow before normal routes.
        roles:
          type: array
          items:
            type: string
        permissions:
          type: array
          items:
            type: string
    UserStatus:
      type: string
      enum:
      - active
      - disabled
      - locked
      description: Account lifecycle status owned by auth.
    ManagedUserRole:
      type: string
      enum:
      - standard
      - admin
      description: Role values assignable through admin user management. super_admin is intentionally excluded.
    UserProfile:
      allOf:
      - $ref: '#/components/schemas/UserSummary'
      - type: object
        required:
        - status
        - mustChangePassword
        properties:
          createdAt:
            type: string
            format: date-time
          updatedAt:
            type: string
            format: date-time
...
    AdminUser:
      allOf:
      - $ref: '#/components/schemas/UserProfile'
      - type: object
        required:
        - createdAt
        - updatedAt
        properties:
          manageableRoles:
            type: array
            items:
              $ref: '#/components/schemas/ManagedUserRole'
            description: Roles the current caller may set for this row. Empty when role change is not allowed.
          actions:
            type: object
            required:
            - canDisable
            - canEnable
            - canResetPassword
            - canChangeRole
            properties:
              canDisable:
                type: boolean
              canEnable:
                type: boolean
              canResetPassword:
                type: boolean
              canChangeRole:
                type: boolean
            description: Stable action hints for management UI rendering; auth remains final authority.
    AdminUserResponse:
      type: object
      required:
      - data
      - requestId
      properties:
        data:
          $ref: '#/components/schemas/AdminUser'
        requestId:
          type: string
    AdminUserListResponse:
      type: object
      required:
      - data
      - page
      - requestId
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AdminUser'
        page:
          $ref: '#/components/schemas/PageInfo'
        requestId:
          type: string
    CreateAdminUserRequest:
      type: object
      required:
      - username
      - temporaryPassword
      - role
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 128
        temporaryPassword:
          type: string
          format: password
          minLength: 8
          maxLength: 1024
          writeOnly: true
          description: Administrator-entered temporary password. It is never returned and forces mustChangePassword.
        role:
          $ref: '#/components/schemas/ManagedUserRole'
        displayName:
          type: string
          maxLength: 128
          description: Optional display name. Empty string is allowed.
        email:
          type: string
          format: email
          nullable: true
          maxLength: 320
          description: Optional non-unique profile email.
        phone:
          type: string
          nullable: true
          maxLength: 32
          description: Optional non-unique profile phone.
...
    SessionSummary:
      type: object
      required:
      - sessionId
      - accessToken
      - tokenType
      - expiresAt
      properties:
        sessionId:
          type: string
        accessToken:
          type: string
          description: Opaque bearer credential sent by the frontend on later requests. This token is not a JWT and must not be parsed for claims.
          example: tok_8Ywq7T2n4pQ9mR3xV6sL
        tokenType:
          type: string
          enum:
          - Bearer
          default: Bearer
        expiresAt:
          type: string
          format: date-time
    SessionResponse:
      type: object
      required:
      - data
      - requestId
      properties:
        data:
          type: object
          required:
          - user
          - session
          properties:
            user:
              $ref: '#/components/schemas/UserSummary'
            session:
              $ref: '#/components/schemas/SessionSummary'
        requestId:
          type: string
...
    PageInfo:
      type: object
      required:
      - page
      - pageSize
      - total
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
...